+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## [Unreleased]
 
+- Fixed default tokenization of heredocs in Ruby (should be tokenized as string)
+- Added a modern implementation of Tree-sitter grammars behind an experimental flag. Enable the “Use Modern Tree-Sitter Implementation” in the Core settings to try it out.
 ## 1.106.0
 
 - Fixed bug that happens on some systems when trying to launch Pulsar using the Cinnamon desktop environment